Rename src/formatstring.cpp after the class it contains
